+++ pkg/webui/console/components/events/messages.js
@@ -32,11 +32,13 @@ const messages = defineMessages({
   rx1Delay: 'Rx1 Delay',
   // Generic messages
   eventDetails: 'Event details',
+  rawEvent: 'Raw event',
   errorOverviewEntry:
     'There was an error and the event cannot be displayed. The raw event can by viewed by clicking this row.',
   dataPreview: 'Data preview',
   seeAllActivity: 'See all activity',
-  syntheticEvent: 'Note: This meta event did not originate from the event stream',
+  syntheticEvent:
+    'Note: This meta event did not originate from the network but was generated automatically by the Console. It is not related to any end device or gateway activity.',
   eventsTruncated:
     'Old events have been truncated to save memory. The current event limit per stream is {limit}.',
   eventUnavailable: 'This event is not available anymore. It was likely truncated to save memory.',